separate filter group into a separate component
